// src/interfaces.ts
export interface teamsDbItem {
    teamname: string;
    teamcaptain: string;
    teamplayer1: string;
    teamplayer2: string;
    teamplayer3: string;
    teamplayer4: string;
    teamplayer5: string;
    teamplayer6: string;
}

export interface orderOfPlayDbItem {
    teamname: string;
    game1Player1: string;
    game1Player2: string;
    game2Player1: string;
    game2Player2: string;
    game3Player1: string;
    game3Player2: string;
    game4Player1: string;
    game4Player2: string;
    game5Player1: string;
    game5Player2: string;
    game6Player1: string;
    game6Player2: string;
    game7Player1: string;
    game7Player2: string;
}